Deduplicate api request with a small mixin
